api: Generate JSON tags
This patch updates the type generation code to support adding
JSON tags to generated type field data.

Please note this also causes all of the types in vim25/types
to now use proper casing for their field when marshaled to JSON.
This is likely a breaking change from current behavior where a
field name emitted to JSON would have retained the casing for the
Go field name, instead of starting with a lowercase character.